refactor(v0.52 phase 5): automation/ → scheduler/ 분리
매핑 매트릭스 §3.4 부분 적용. cron/scheduling 관련 5 파일을
core/scheduler/ 로 분리하고, 자율 학습 관련 7 파일은 core/automation/ 잔류.

분류:
core/scheduler/ (cron/job 호스팅)
- scheduler.py        — Scheduler core, ScheduledJob/Schedule
- triggers.py         — TriggerManager (F1-F4 trigger types)
- predefined.py       — pre-defined automation templates
- nl_scheduler.py     — natural language → ScheduledJob 파서
- calendar_bridge.py  — calendar ↔ scheduler 양방향 sync

core/automation/ (자율 학습 잔류)
- correlation.py, drift.py, feedback_loop.py, outcome_tracking.py,
  snapshot.py, expert_panel.py, model_registry.py

config.py 분할은 ROI 낮아 보류 (522줄 단일 파일이지만 잘 동작 중).
Phase 5b 로 분리하거나 필요 시점에 재검토.

Import 경로 갱신: 21 파일 (sed 자동)

Invariant test 갱신:
- DAEMON_DIRS 에 core/scheduler/ 추가

검증:
- ruff + format clean (6 import 자동수정)
- mypy strict clean (231 modules, +1)
- pytest 4081 passed (1 skipped, 19 deselected)
- E2E `geode analyze "Cowboy Bebop" --dry-run` → A (68.4) unchanged

다음 phase: import-linter 도입 + agent/ 내부 rename
