+++ core/scheduler/calendar_bridge.py
