+++ core/tools/calendar_tools.py
@@ -231,7 +231,7 @@ class CalendarSyncSchedulerTool:
         }
 
     def execute(self, **kwargs: Any) -> dict[str, Any]:
-        from core.automation.calendar_bridge import get_calendar_bridge
+        from core.scheduler.calendar_bridge import get_calendar_bridge
 
         bridge = get_calendar_bridge()
         if bridge is None:
