+++ tests/test_scheduler_lock.py
@@ -8,7 +8,7 @@ import time
 from pathlib import Path
 
 import pytest
-from core.automation.scheduler import SchedulerLock, _is_pid_alive
+from core.scheduler.scheduler import SchedulerLock, _is_pid_alive
 
 
 class TestSchedulerLock:
