+++ tests/test_scheduler_missed.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import time
 from pathlib import Path
 
-from core.automation.scheduler import (
+from core.scheduler.scheduler import (
     MISSED_TASK_GRACE_MS,
     Schedule,
     ScheduledJob,
